+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+# 1.3.0
+- now you can add style for button elements separately.
+- regex update for better matching color values like rgb, rgba, linear-gradient
+- dependencies updates
+	- postcss 5.0.21 => 5.2.6
+	- ava 0.16.0 => 0.17.0
+	- eslint 3.1.1 => 3.11.0
+- tests updates
+- minor fixes
 # 1.2.0
 - remove css props option
 - selector list function fix